+++ tests/testthat/test-ggmice.R
@@ -11,6 +11,8 @@ imp <- suppressWarnings(mice::mice(dat, printFlag = FALSE))
 test_that("continuous data plot", {
   expect_s3_class(ggmice(dat, ggplot2::aes(bmi)), "ggplot")
   expect_s3_class(ggmice(imp, ggplot2::aes(bmi)), "ggplot")
+  expect_s3_class(ggmice(dat, ggplot2::aes(y = bmi)), "ggplot")
+  expect_s3_class(ggmice(imp, ggplot2::aes(y = bmi)), "ggplot")
 })
 
 test_that("categorical data plot", {
@@ -31,7 +33,8 @@ test_that("mixed data plot", {
 
 test_that("complete data plot", {
   expect_s3_class(ggmice(na.omit(dat), ggplot2::aes(bmi)), "ggplot")
-  imp2 <- suppressWarnings(mice::mice(na.omit(dat), printFlag = FALSE, seed = 1))
+  imp2 <-
+    suppressWarnings(mice::mice(na.omit(dat), printFlag = FALSE, seed = 1))
   expect_s3_class(ggmice(imp2, ggplot2::aes(bmi)), "ggplot")
 })
 
@@ -42,11 +45,15 @@ test_that("incorrect data", {
   expect_error(ggmice(cbind(dat, dat), ggplot2::aes(bmi)))
 })
 
+test_that("advanced mapping", {
+  expect_error(ggmice(dat, ggplot2::aes(log(age))))
+  expect_error(ggmice(dat, ggplot2::aes(age3)))
+  expect_warning(ggmice(dat, ggplot2::aes(bmi, color = bmi)))
+})
+
 test_that("incorrect mapping", {
   expect_error(ggmice(dat))
   expect_error(ggmice(dat, ggplot2::aes(x = test)))
   expect_error(ggmice(dat, ggplot2::aes(group = age)))
   expect_error(ggmice(dat, ggplot2::aes("bmi")))
-  expect_warning(ggmice(dat, ggplot2::aes(bmi, color = bmi)))
 })
-